requirements.txt
absl-py==2.0.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.30.2
azure-storage-blob==12.16.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blis==0.7.11
boto3==1.34.130
botocore==1.34.130
CacheControl==0.12.14
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==2.0.0
confection==0.1.5
confluent-kafka==2.4.0
contourpy==1.2.1
cryptography==42.0.8
cvg-python-sdk==0.5.1
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
environs==9.5.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
Faker==19.13.0
fbmessenger==6.0.0
filelock==3.15.3
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
gast==0.4.0
google-api-core==2.8.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.56.1
greenlet==3.0.3
grpcio==1.63.0
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hvac==1.2.1
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
isodate==0.6.1
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.2.0
jq==1.7.0
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
keras==2.12.0
kiwisolver==1.4.5
langchain==0.0.329
langcodes==3.4.0
langsmith==0.0.92
language_data==1.2.0
libclang==18.1.1
llama-index==0.10.47
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.47
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
locket==1.0.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5
mattermostwrapper==2.2
mdurl==0.1.2
milvus-lite==2.4.7
ml-dtypes==0.4.0
msgpack==1.0.8
multidict==5.2.0
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
ollama==0.2.1
ollama-python==0.1.2
openai==1.35.1
opentelemetry-api==1.15.0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==2.2.2
partd==1.4.2
pep440-version-utils==0.3.0
phonenumbers==8.13.39
pillow==10.3.0
pluggy==1.5.0
ply==3.11
portalocker==2.8.2
preshed==3.0.9
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
prompt-toolkit==3.0.28
protobuf==4.23.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydot==1.4.2
Pygments==2.18.0
pyhcl==0.4.5
PyJWT==2.8.0
pykwalify==1.8.0
pymilvus==2.4.3
pymongo==4.3.3
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypred==0.4.0
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-keycloak==3.12.0
python-socketio==5.11.3
pytz==2022.7.1
PyYAML==6.0.1
qdrant-client==1.9.1
questionary==1.10.0
randomname==0.2.1
rasa==3.6.20
rasa-pro==3.8.0
rasa-sdk==3.8.0
redis==4.6.0
referencing==0.35.1
regex==2022.10.31
reportlab==4.2.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.24.1
rich==13.7.1
rocketchat-API==1.30.0
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safetensors==0.4.3
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.29.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
striprtf==0.0.26
structlog==23.1.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.5
tenacity==8.4.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.2.5
threadpoolctl==3.5.0
thrift==0.20.0
tiktoken==0.4.0
tldextract==5.1.2
tokenizers==0.19.1
toolz==0.12.1
tqdm==4.66.4
transformers==4.41.2
twilio==8.4.0
typer==0.12.3
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uv==0.2.13
uvloop==0.19.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.3
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
